% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{images2matrix}
\alias{images2matrix}
\title{Estimate number of chunks based on object size}
\usage{
images2matrix(image_list, mask)
}
\arguments{
\item{image_list}{list containing image paths.}

\item{mask}{mask image paths.}
}
\value{
Returns matrix of dimension (numImages, numVoxelsInMask)
}
\description{
Estimate number of chunks based on object size
}
