% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support.R
\name{james_stein}
\alias{james_stein}
\title{Apply kernel james-stein shrinkage to intercept coefficients of a vbgamlss}
\usage{
james_stein(
  object,
  mask,
  yimageframe,
  save_model = NULL,
  radius_voxels = 4,
  num_cores = NULL
)
}
\description{
Apply kernel james-stein shrinkage to intercept coefficients of a vbgamlss
}
