% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{load_input_image}
\alias{load_input_image}
\title{Check input data}
\usage{
load_input_image(image, mask = NULL)
}
\arguments{
\item{image}{path of the image to load and check. RDS, nifti or dataframe}

\item{mask}{path to mask.}
}
\value{
image as a dataframe.
}
\description{
Check input data
}
