% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support.R
\name{load_model}
\alias{load_model}
\title{Load VBGAMLSS models from files with the specified prefix.}
\usage{
load_model(file_prefix, num_cores = NULL)
}
\arguments{
\item{file_prefix}{Prefix of the files containing the VBGAMLSS models.}

\item{num_cores}{Number of CPU cores to use for parallel processing.
Defaults to all available cores if not provided.}
}
\value{
A structure containing the loaded VBGAMLSS models.
}
\description{
Load VBGAMLSS models from files with the specified prefix.
}
