% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mapping.R
\name{map_model_predictions}
\alias{map_model_predictions}
\title{Write prediction maps to NIfTI files}
\usage{
map_model_predictions(obj, mask, filename, index = NULL, return_files = FALSE)
}
\arguments{
\item{obj}{A \code{"vbgamlss.predictions"} object. Each voxel entry should  contain \code{$family}, optional \code{$vxl}, and one element per modelparameter (for example \code{mu}, \code{sigma}), each being a numeric vector over subjects with consistent ordering across voxels.}

\item{mask}{An ANTs image (or path) providing geometry, used to map matrices back to images.}

\item{filename}{Character prefix for output files. The function appends \code{"_subj-<ID>_fam-<FAMILY>_par-<PARAM>.nii.gz"}.}

\item{index}{Optional integer vector of subject indices to map. If \code{NULL} all subjects are processed.}

\item{return_files}{Logical, if \code{TRUE} return the vector of file paths written for the last parameter processed. Default \code{FALSE}.}
}
\description{
Convert per-voxel predicted parameter (μ,σ,ν,τ) values from a \code{vbgamlss.predictions} object into images and save them to disk.
}
