% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support.R
\name{predict.vbgamlss}
\alias{predict.vbgamlss}
\title{Predict method for vbgamlss objects}
\usage{
\method{predict}{vbgamlss}(
  object,
  newdata = NULL,
  num_cores = NULL,
  ptype = "parameter",
  segmentation = NULL,
  segmentation_target = NULL,
  terms = NULL,
  what = NULL,
  future_plan_strategy = "future.mirai::mirai_cluster",
  chunk_max_mb = 1024,
  se.fit = FALSE,
  ...
)
}
\arguments{
\item{object}{vbgamlss object to make predictions from.}

\item{newdata}{New data to use for predictions. Must be data.frame.}

\item{num_cores}{Number of CPU cores to use for parallel processing.
Defaults to available cores if not provided.}

\item{ptype}{Type of prediction to make: "parameter", "link", "response", "terms".
Defaults to "parameter".}

\item{segmentation}{Optional data.frame or matrix with region labels (e.g., tissue types)
mapping to the voxels in the object.}

\item{segmentation_target}{Optional integer. If provided, the prediction is only
performed for voxels where the segmentation matches this value.}

\item{terms}{Optional character vector. What terms to include in prediction. See gamlss2 doc.}

\item{what}{Optional character. Specifies which distribution parameter to predict
(e.g., "mu", "sigma").}

\item{future_plan_strategy}{Character. The parallel back-end strategy to use.
Defaults to "future.mirai::mirai_cluster".}

\item{chunk_max_mb}{Integer. Maximum size in MB for each data chunk sent to workers.
Defaults to 1024.}

\item{...}{Additional arguments passed to the underlying predict.gamlss2 function.}
}
\value{
A structure of class "vbgamlss.predictions" containing a list of
prediction results for each voxel.
}
\description{
Predict method for vbgamlss objects
}
