% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{restore_family}
\alias{restore_family}
\title{Wrapper to reassign removed family() object from a vbgamlss fitted submodel}
\usage{
restore_family(fitted)
}
\arguments{
\item{fitted, }{vbgamlss fitted submodel,}
}
\value{
vbgamlss fitted submodel with family object included.
}
\description{
Wrapper to reassign removed family() object from a vbgamlss fitted submodel
}
