% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support.R
\name{save_model}
\alias{save_model}
\title{Save VBGAMLSS models from files with the specified prefix.}
\usage{
save_model(model_list, filename, voxel = NULL)
}
\arguments{
\item{model_list}{vbgamlss fitted model.}

\item{filename}{Prefix of the files to save the VBGAMLSS models.}

\item{voxel}{Index/s of the Voxel/s to save model for. If NULL, save the entire
model.}
}
\description{
Save VBGAMLSS models from files with the specified prefix.
}
