% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core.R
\name{[[.vbgamlss}
\alias{[[.vbgamlss}
\title{Define the S3 subsetting method, here it would cool to include restore_family..}
\usage{
\method{[[}{vbgamlss}(x, i, ...)
}
\description{
Define the S3 subsetting method, here it would cool to include restore_family..
}
