% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{vbapply}
\alias{vbapply}
\title{Wrapper to apply a function to all submodels of a vbgamlss fitted object}
\usage{
vbapply(fitted, AFUN, num_cores = 1L, ...)
}
\arguments{
\item{fitted}{A vbgamlss fitted model (list of submodels).}

\item{AFUN}{The function to apply. eg. AIC(), kurtosis(), etc.}

\item{num_cores}{Integer. Number of cores for parallel processing. Defaults to 1.}

\item{...}{Additional arguments passed to FUN.}
}
\value{
A list of outputs.
}
\description{
Wrapper to apply a function to all submodels of a vbgamlss fitted object
}
