% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support.R
\name{zscore.vbgamlss}
\alias{zscore.vbgamlss}
\title{Compute Z-scores for vbgamlss predictions given Y voxel data and image mask.}
\usage{
zscore.vbgamlss(predictions, yimageframe, num_cores = NULL)
}
\arguments{
\item{predictions}{A vbgamlss.predictions object.}

\item{yimageframe}{A data.frame or matrix of observed response values.}

\item{num_cores}{Number of CPU cores to use for parallel processing.}
}
\value{
A structure containing z-scores.
}
\description{
Compute Z-scores for vbgamlss predictions given Y voxel data and image mask.
}
