% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support.R
\name{load_model}
\alias{load_model}
\title{Load VBGAMLSS models from files with the specified prefix.}
\usage{
load_model(filepath)
}
\arguments{
\item{filepath}{path to serialized model.}
}
\value{
A structure containing the loaded VBGAMLSS models.
}
\description{
Load VBGAMLSS models from files with the specified prefix.
}
