% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support.R
\name{predict.vbgamlss}
\alias{predict.vbgamlss}
\title{Predict method for vbgamlss objects}
\usage{
\method{predict}{vbgamlss}(object, newdata = NULL, num_cores = NULL, ...)
}
\arguments{
\item{object}{vbgamlss object to make predictions from.}

\item{newdata}{New data to use for predictions.}

\item{num_cores}{Number of CPU cores to use for parallel processing.
Defaults to one less than the total available cores if not provided.}

\item{...}{Additional arguments to be passed to the \code{\link{predict}} function.}
}
\value{
A structure containing predictions.
}
\description{
Predict method for vbgamlss objects
}
