% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core.R
\name{vbgamlss}
\alias{vbgamlss}
\title{Fit Generalized Additive Models for Location Scale and Shape (GAMLSS) with Voxel Data}
\usage{
vbgamlss(
  imageframe,
  g.formula,
  train.data,
  g.family = NO,
  segmentation = NULL,
  segmentation_target = NULL,
  num_cores = NULL,
  chunk_max_mb = 64,
  debug = F,
  cachedir = getwd(),
  force_constraints = c(1e-08, +Inf),
  warm_start = NULL,
  show_progress = T,
  eps = 1e-05,
  maxit = c(100, 33),
  future_plan_strategy = "future.mirai::mirai_cluster",
  save_model = NULL,
  deep_stripping = TRUE,
  ...
)
}
\arguments{
\item{imageframe}{A dataframe containing the voxel data. Each column represents a voxel/vertex,
and each row represents a subject.}

\item{g.formula}{A formula for the GAMLSS2 model.}

\item{train.data}{A data frame containing subject-level data. Columns must correspond to covariates in \code{g.formula}.}

\item{g.family}{A GAMLSS family object. Defaults to \code{NO} (Normal).}

\item{segmentation}{Optional, dataframe generated from a mask containing segmentation information (e.g. 1=GM, 2=WM). It must have the shape as \code{imageframe}.}

\item{segmentation_target}{Optional, integer, target/label for the segmentation data to subset the analysis (e.g. 1 for GM).}

\item{num_cores}{Number of cores for parallel processing. If NULL, all available cores are used.}

\item{chunk_max_mb}{Maximum chunk size in megabytes for processing. Defaults to 64 MB. Increase to 128/256/512 for HPC.}

\item{debug}{Logical. If TRUE, enables debug mode to log output in \code{logdir}. Defaults to FALSE.}

\item{cachedir}{Character, path for cached artifacts.}

\item{force_constraints}{list, force voxels outside constraints to be excluded from the fit:  Constr1 <= voxels <= Constr2. Use NULL to exclude constraints. Defaults to c(1e-8, +Inf)}

\item{...}{Additional arguments passed to the GAMLSS fitting function.}

\item{afold}{Optional boolean or integer vector to subset imageframe for cross-validation folds.}

\item{logdir}{Directory path for saving logs if \code{debug} is TRUE. Defaults to current working directory.}

\item{cache}{Logical, cache intermediate results.}
}
\value{
A list of GAMLSS models, with class \code{"vbgamlss"}.
}
\description{
The \code{vbgamlss} function fits Bayesian GAMLSS models on voxel data with optional segmentation,
allowing for parallel processing. This function can be used to analyze high-dimensional imaging data.
}
\details{
The function performs the following steps:
\itemize{
\item Checks input data and prepares voxel and segmentation data as needed.
\item If parallel processing is enabled, splits voxel data into chunks based on memory constraints.
Each chunks are processed sequentially, but voxels/vertex within each chunk are processed in parallel.
\item Fits a GAMLSS model to each voxel using specified covariates and segmentation data, if available.
\item Returns a vbgamlss object containing the list of models, one for each voxel.
}
}
