% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core.R
\name{vbgamlss}
\alias{vbgamlss}
\title{Fit Generalized Additive Models for Location Scale and Shape (GAMLSS) with Voxel Data}
\usage{
vbgamlss(
  imageframe,
  g.formula,
  train.data,
  g.family = NO,
  segmentation = NULL,
  segmentation_target = NULL,
  num_cores = NULL,
  chunk_max_mb = 64,
  afold = NULL,
  subsample = NULL,
  debug = F,
  logdir = getwd(),
  cache = F,
  cachedir = getwd(),
  force_ypositivity = T,
  ...
)
}
\arguments{
\item{imageframe}{A data frame containing the voxel data. Each column represents a voxel,
and each row represents a subject.}

\item{g.formula}{A formula for the GAMLSS model.}

\item{train.data}{A data frame containing subject-level data. Columns must correspond to covariates in \code{g.formula}.}

\item{g.family}{A GAMLSS family object. Defaults to \code{NO} (Normal).}

\item{segmentation}{Optional data frame containing segmentation information.}

\item{segmentation_target}{Optional target for the segmentation data to subset the analysis.}

\item{num_cores}{Number of cores for parallel processing. If NULL, all available cores are used.}

\item{chunk_max_mb}{Maximum chunk size in megabytes for processing. Defaults to 64 MB. Increase to 256/512 for HPC.}

\item{afold}{Optional boolean or integer vector to subset imageframe for cross-validation folds.}

\item{subsample}{(to be DEPRECATED) Optional numeric vector of voxel indices to subset the imageframe.}

\item{debug}{Logical. If TRUE, enables debug mode to log output in \code{logdir}. Defaults to FALSE.}

\item{logdir}{Directory path for saving logs if \code{debug} is TRUE. Defaults to current working directory.}

\item{cache}{Logical, cache intermediate results.}

\item{cachedir}{Character path for cached artifacts.}

\item{force_ypositivity}{Logical, shift/clip y to be positive if the family requires it.}

\item{...}{Additional arguments passed to the GAMLSS fitting function.}
}
\value{
A list of GAMLSS models, with class \code{"vbgamlss"}.
}
\description{
The \code{vbgamlss} function fits Bayesian GAMLSS models on voxel data with optional segmentation,
allowing for parallel processing. This function can be used to analyze high-dimensional imaging data.
}
\details{
The function performs the following steps:
\itemize{
\item Checks input data and prepares voxel and segmentation data as needed.
\item If parallel processing is enabled, splits voxel data into chunks based on memory constraints.
\item Fits a GAMLSS model to each voxel using specified covariates and segmentation data, if available.
\item Returns a list of models, one for each voxel.
}
}
